using double instead of float

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/localization/Position.java
@@ -1,5 +1,5 @@
 package org.firstinspires.ftc.teamcode.autonomous.localization;
 
 public class Position {
-    public Float x, y, t;
+    public double x, y, t;
 }
